+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/ErrorResponse.kt
@@ -34,6 +34,15 @@ sealed interface ErrorResponse {
     ) : ErrorResponse
 
     /**
+     * The account has been suspended and can only be used for limited actions at this time.
+     */
+    @Serializable
+    @SerialName("M_USER_SUSPENDED")
+    data class UserSuspended(
+        override val error: String,
+    ) : ErrorResponse
+
+    /**
      * The account has been locked and cannot be used at this time.
      */
     @Serializable
@@ -144,6 +153,13 @@ sealed interface ErrorResponse {
     @Serializable
     @SerialName("M_THREEPID_IN_USE")
     data class ThirdPIdInUse(override val error: String) : ErrorResponse
+
+    /**
+     * The homeserver does not support adding a third party identifier of the given medium.
+     */
+    @Serializable
+    @SerialName("M_THREEPID_MEDIUM_NOT_SUPPORTED")
+    data class ThirdPartyMediumNotSupported(override val error: String) : ErrorResponse
 
     /**
      * Sent when a threepid given to an API cannot be used because no record matching the threepid was found.