+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Keep isDirect for room with membership leave or ban.
+
 ### Deprecated
 
 ### Removed